// src/models/pbdModuleString.ts
export enum PbdModuleString {
  None = "None",
  DocumentManagement = "DocumentManagement",
  GoalManagement = "GoalManagement",
  OpportunityManagement = "OpportunityManagement",
  Settings = "Settings",
  KnowledgeBase = "KnowledgeBase",
  QualityMonitor = "QualityMonitor",
  ContactManagement = "ContactManagement",
  Dev = 8,
  Blog = "Blog",
  Help = "Help",
  ToDoManagement = "ToDoManagement",
  ClaimManagement = "ClaimManagement",
  TrainingManagement = "TrainingManagement",
  MaintenanceManagement = "MaintenanceManagement",
  AuditManagement = "AuditManagement",
  Article = 16,
  ProjectManagement = "ProjectManagement",
  DefectManagement = "DefectManagement",
  IdeaManagement = "IdeaManagement",
  IdeaManagementManager = 20,
  Products = "Products",
  TenantManagement = "TenantManagement",
  IdeaManagementCoordinator = 23,
  CustomerSatisfactionManagement = "CustomerSatisfactionManagement",
  SupplierSatisfactionManagement = "SupplierSatisfactionManagement",
  Employees = "Employees",
  QualificationMatrix = "QualificationMatrix",
  ProjectAndTaskManagement = "ProjectAndTaskManagement",
  Admin = "Admin",
  Organisations = "Organisations",
  CustomForms = "CustomForms",
  Startpage = "Startpage",
  CompanyFunction = "CompanyFunction",
  Qualification = "Qualification",
  TenantQualificationDone = "TenantQualificationDone",
  CompanyFunctionMatrix = "CompanyFunctionMatrix",
  Crm = "Crm",
  AbsencePlanner = "AbsencePlanner",
  Store = "Store",
}